Refactor unet and unetr benchmarking scripts
